[BOOK-161] refactor: Search에 공통 컴포넌트 적용
