+++ feature/search/src/main/kotlin/com/ninecraft/booket/feature/search/SearchPresenter.kt
@@ -12,6 +12,7 @@ import com.ninecraft.booket.core.common.utils.handleException
 import com.ninecraft.booket.core.data.api.repository.BookRepository
 import com.ninecraft.booket.core.model.BookSearchModel
 import com.ninecraft.booket.core.model.BookSummaryModel
+import com.ninecraft.booket.core.ui.component.FooterState
 import com.ninecraft.booket.feature.screens.LoginScreen
 import com.ninecraft.booket.feature.screens.SearchScreen
 import com.orhanobut.logger.Logger
